refactor: move request() from index.js to util.js
